# docs/03-API/openapi-mvp-v1-consolidated.yaml
openapi: 3.1.0
info:
  title: TransportERP-UA API
  version: 1.0.0-m0
  description: Машинозчитуваний контракт першого production MVP. Канонічний опис архітектури та бізнес-семантики
    ведеться українською у docs/03-API.
servers:
- url: /api/v1
security:
- sessionAuth: []
...
paths:
  /auth/me:
    get:
      tags:
      - Identity
      summary: Поточний користувач та effective permissions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
          description: Session required
      security:
      - SessionCookie: []
...
  /auth/login:
    post:
      tags:
      - Identity
      summary: Створити browser session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Session created; opaque token is returned only as HttpOnly cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          description: Invalid credentials
        '403':
          description: Account unavailable
  /auth/refresh:
    post:
      tags:
      - Identity
      summary: Rotate current browser session
      security:
      - SessionCookie: []
        CsrfHeader: []
      responses:
        '200':
          description: Session rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          description: Session unavailable
  /auth/logout:
    post:
      tags:
      - Identity
      summary: Revoke current browser session
      security:
      - SessionCookie: []
        CsrfHeader: []
      responses:
        '204':
          description: Session revoked
  /auth/permissions:
    get:
      tags:
      - Identity
      summary: Effective permissions поточного користувача
      security:
      - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'
  /users:
    get:
      tags:
      - Identity
      summary: Користувачі tenant
      x-permission: user.read
      security:
      - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityUserListResponse'
    post:
      tags:
      - Identity
      summary: Створити користувача tenant
      x-permission: user.create
      security:
      - SessionCookie: []
        CsrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityUserCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityUserResponse'
  /users/{user_id}:
    parameters:
    - name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      tags:
      - Identity
      summary: Користувач tenant
      x-permission: user.read
      security:
      - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityUserResponse'
    patch:
      tags:
      - Identity
      summary: Оновити profile fields користувача
      x-permission: user.update
      security:
      - SessionCookie: []
        CsrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityUserPatch'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityUserResponse'
  /users/{user_id}/activate:
    post:
      tags:
      - Identity
      summary: Активувати користувача
      x-permission: user.status.change
      security:
      - SessionCookie: []
        CsrfHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Activated
  /users/{user_id}/suspend:
    post:
      tags:
      - Identity
      summary: Призупинити користувача
      x-permission: user.status.change
      security:
      - SessionCookie: []
        CsrfHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Suspended
  /users/{user_id}/roles:
    parameters:
    - name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      tags:
      - Identity
      summary: Ролі користувача
      x-permission: user.read
      security:
      - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRolesResponse'
    put:
      tags:
      - Identity
      summary: Замінити набір ролей користувача
      x-permission: user.roles.manage
      security:
      - SessionCookie: []
        CsrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRolesReplace'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRolesResponse'
  /roles:
    get:
      tags:
      - Identity
      summary: Системні та tenant roles, видимі поточному tenant
      x-permission: role.read
      security:
      - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityRoleListResponse'
    post:
      tags:
      - Identity
      summary: Створити tenant role
      x-permission: role.manage
      security:
      - SessionCookie: []
        CsrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityRoleCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityRoleResponse'
  /permissions:
    get:
      tags:
      - Identity
      summary: Каталог stable permission codes
      x-permission: permission.read
      security:
      - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionListResponse'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Secure HttpOnly browser session cookie. Exact provider may evolve without changing
        domain API.
    SessionCookie:
      type: apiKey
      in: cookie
      name: transport_erp_session
    CsrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  responses:
    Unauthenticated:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Meta:
      type: object
      required:
      - request_id
      properties:
        request_id:
          type: string
          format: uuid
...
    ErrorResponse:
      type: object
      required:
      - error
      - meta
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          $ref: '#/components/schemas/Meta'
    AuthMeResponse:
      type: object
      required:
      - data
      properties:
        data:
          type: object
          required:
          - id
          - username
          - permissions
          - locale
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            permissions:
              type: array
              items:
                type: string
            locale:
              type: string
              enum:
              - uk
              - en
              - es
              - fr
              - de
          $ref: '#/components/schemas/AuthMeData'
        meta:
          $ref: '#/components/schemas/Meta'
...
    IdentityUser:
      type: object
      required:
      - id
      - company_id
      - username
      - status
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - DISABLED
        preferred_locale:
          type:
          - string
          - 'null'
          enum:
          - uk
          - en
          - es
          - fr
          - de
          - null
    AuthLoginRequest:
      type: object
      required:
      - company_edrpou
      - username
      - password
      properties:
        company_edrpou:
          type: string
          pattern: ^[0-9]{8,10}$
        username:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 256
          writeOnly: true
    AuthMeData:
      type: object
      required:
      - user
      - permissions
      properties:
        user:
          $ref: '#/components/schemas/IdentityUser'
        permissions:
          type: array
          items:
            type: string
    PermissionListResponse:
      type: object
      required:
      - data
      properties:
        data:
          type: array
          items:
            type: string
    IdentityUserCreate:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type:
          - string
          - 'null'
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 256
          writeOnly: true
        preferred_locale:
          type:
          - string
          - 'null'
          enum:
          - uk
          - en
          - es
          - fr
          - de
          - null
    IdentityUserPatch:
      type: object
      properties:
        email:
          type:
          - string
          - 'null'
          maxLength: 320
        preferred_locale:
          type:
          - string
          - 'null'
          enum:
          - uk
          - en
          - es
          - fr
          - de
          - null
    IdentityUserResponse:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/IdentityUser'
    IdentityUserListResponse:
      type: object
      required:
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/IdentityUser'
    IdentityRole:
      type: object
      required:
      - id
      - code
      - name
      - system_role
      - active
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type:
          - string
          - 'null'
          format: uuid
        code:
          type: string
        name:
          type: string
        system_role:
          type: boolean
        active:
          type: boolean
    IdentityRoleCreate:
      type: object
      required:
      - code
      - name
      properties:
        code:
          type: string
          minLength: 2
          maxLength: 80
        name:
          type: string
          minLength: 1
          maxLength: 200
        permissions:
          type: array
          items:
            type: string
          default: []
    IdentityRoleResponse:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/IdentityRole'
    IdentityRoleListResponse:
      type: object
      required:
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/IdentityRole'
    UserRolesReplace:
      type: object
      required:
      - role_ids
      properties:
        role_ids:
          type: array
          items:
            type: string
            format: uuid
    UserRolesData:
      type: object
      required:
      - user_id
      - roles
      properties:
        user_id:
          type: string
          format: uuid
        roles:
          type: array
          items:
            type: string
    UserRolesResponse:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/UserRolesData'